HelloDjango/buttons/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from .models import Book, Genre

class BookSerializer(serializers.ModelSerializer):

    class Meta:
        model = Book
        fields = '__all__'


class GenreSerializer(serializers.ModelSerializer):

    class Meta:
        model = Genre
        fields = '__all__'